+++ app/config/settings/production.py
@@ -7,3 +7,12 @@ ALLOWED_HOSTS = [
 ]
 
 WSGI_APPLICATION = 'config.wsgi.production.application'
+
+sentry_sdk.init(
+    dsn=SECRETS['dsn'],
+    integrations=[DjangoIntegration()],
+
+    # If you wish to associate users to errors (assuming you are using
+    # django.contrib.auth) you may enable sending PII data.
+    send_default_pii=True
+)
