+++ app/contents/serializers.py
@@ -1,8 +1,7 @@
 from rest_framework import serializers
 
 from contents.models import Contents, Video
-from members.models import Watching
-from members.serializers import ProfileSerializer
+from members.models import Watching, Profile
 
 
 class ContentsDetailSerializer(serializers.ModelSerializer):
@@ -21,9 +20,20 @@ class ContentsDetailSerializer(serializers.ModelSerializer):
             'contents_rating',
             'contents_length',
             'contents_pub_year',
+            'preview_video',
             'actors',
             'directors',
         ]
+
+    def to_representation(self, instance):
+        profile = Profile.objects.get(pk=self.context['profile_pk'])
+        is_selected = True if profile in instance.select_profiles.all() else False
+        is_like = True if profile in instance.like_profiles.all() else False
+
+        representation = super().to_representation(instance)
+        representation['is_selected'] = is_selected
+        representation['is_like'] = is_like
+        return representation
 
 
 class ContentsSerializer(serializers.ModelSerializer):
